+++ SignNow.Net/Model/SignNowDocument.cs
@@ -10,14 +10,14 @@ namespace SignNow.Net.Model
 {
     /// <summary>
     /// Represents SignNow document object.
-    /// <remarks>
-    /// Document is the fundamental unit of every e-Signature operation. It contains:
-    ///     <para>Metadata: file name, size, extension, ID;</para>
-    ///     <para>Fields, elements (texts, checks, signatures, etc...);</para>
-    ///     <para>Invites, statuses of the invites;</para>
-    ///     <para>Document <see cref="Role"/>;</para>
-    ///     <para>Timestamps (date created, date updated);</para>
-    /// </remarks>
+    /// Document is the fundamental unit of every e-Signature operation.
+    /// <para>It contains:</para>
+    /// <list type="bullet">
+    /// <item><description>Metadata: file name, size, extension, ID;</description></item>
+    /// <item><description>Fields, field content elements (texts, checks, signatures, etc...);</description></item>
+    /// <item><description>Invites, statuses of the invites;</description></item>
+    /// <item><description>Document <see cref="Role"/>;</description></item>
+    /// </list>
     /// </summary>
     public partial class SignNowDocument
     {
@@ -94,19 +94,25 @@ namespace SignNow.Net.Model
         /// The document signer roles.
         /// </summary>
         [JsonProperty("roles")]
-        internal List<Role> Roles { get; private set; } = new List<Role>();
+        internal List<Role> Roles { get; set; } = new List<Role>();
 
         /// <summary>
         /// The document <see cref="SignatureContent"/>
         /// </summary>
         [JsonProperty("signatures")]
-        internal List<SignatureContent> Signatures { get; private set; } = new List<SignatureContent>();
+        internal List<SignatureContent> Signatures { get; set; } = new List<SignatureContent>();
 
         /// <summary>
         /// The document <see cref="Field"/>
         /// </summary>
         [JsonProperty("fields")]
-        public IReadOnlyCollection<Field> Fields { get; private set; } = new List<Field>();
+        internal List<Field> fields { get; set; } = new List<Field>();
+
+        /// <summary>
+        /// All the document fields
+        /// </summary>
+        [JsonIgnore]
+        public IReadOnlyCollection<ISignNowField> Fields => fields;
 
         /// <summary>
         /// The document freeform invite requests.